feat(spark): Add index lambda support for Spark filter function
Spark's filter(array, func) supports two lambda forms:
1. filter(array, x -> predicate) - element only
2. filter(array, (x, i) -> predicate) - element + 0-based index

The current implementation reuses Presto's udf_array_filter which only
supports form (1). This adds a Spark-specific FilterFunction that
supports both forms, following the same pattern as Transform.cpp.

Changes:
- Add FilterFunction.cpp with both 1-arg and 2-arg lambda signatures
- Register udf_spark_filter in sparksql namespace (replacing udf_array_filter)
- Add FilterTest.cpp with 6 test cases covering element-only, element+index,
  index-only, nulls, and isnotnull predicates
