+++ velox/functions/prestosql/FilterFunctions.cpp
@@ -13,161 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "velox/expression/Expr.h"
 #include "velox/expression/LambdaExpr.h"
-#include "velox/expression/VectorFunction.h"
-#include "velox/functions/lib/LambdaFunctionUtil.h"
-#include "velox/functions/lib/RowsTranslationUtil.h"
+#include "velox/functions/lib/FilterFunctionBase.h"
 #include "velox/vector/FlatMapVector.h"
-#include "velox/vector/FunctionVector.h"
 
 namespace facebook::velox::functions {
 namespace {
-
-class FilterFunctionBase : public exec::VectorFunction {
- protected:
-  // Applies filter functions to elements of maps or arrays and returns the
-  // number of elements that passed the filters. Stores the number of elements
-  // in each array or map that passed the filter in resultSizes. Stores the
-  // indices of passing elements in selectedIndices. resultOffsets is populated
-  // assuming sequential layout of the passing elements.
-  template <typename T>
-  static vector_size_t doApply(
-      const SelectivityVector& rows,
-      const std::shared_ptr<T>& input,
-      const VectorPtr& lambda,
-      const std::vector<VectorPtr>& lambdaArgs,
-      exec::EvalCtx& context,
-      BufferPtr& resultOffsets,
-      BufferPtr& resultSizes,
-      BufferPtr& selectedIndices) {
-    const auto* inputOffsets = input->rawOffsets();
-    const auto* inputSizes = input->rawSizes();
-
-    auto* pool = context.pool();
-    resultSizes = allocateSizes(rows.end(), pool);
-    resultOffsets = allocateOffsets(rows.end(), pool);
-    auto* rawResultSizes = resultSizes->asMutable<vector_size_t>();
-    auto* rawResultOffsets = resultOffsets->asMutable<vector_size_t>();
-
-    const auto numElements = lambdaArgs[0]->size();
-
-    selectedIndices = allocateIndices(numElements, pool);
-    auto* rawSelectedIndices = selectedIndices->asMutable<vector_size_t>();
-
-    vector_size_t numSelected = 0;
-
-    auto elementToTopLevelRows =
-        getElementToTopLevelRows(numElements, rows, input.get(), pool);
-
-    exec::LocalDecodedVector bitsDecoder(context);
-    auto iter = lambda->asUnchecked<FunctionVector>()->iterator(&rows);
-    while (auto entry = iter.next()) {
-      auto elementRows =
-          toElementRows<T>(numElements, *entry.rows, input.get());
-      auto wrapCapture =
-          toWrapCapture<T>(numElements, entry.callable, *entry.rows, input);
-
-      VectorPtr bits;
-      entry.callable->apply(
-          elementRows,
-          nullptr,
-          wrapCapture,
-          &context,
-          lambdaArgs,
-          elementToTopLevelRows,
-          &bits);
-      bitsDecoder.get()->decode(*bits, elementRows);
-      entry.rows->applyToSelected([&](vector_size_t row) {
-        if (input->isNullAt(row)) {
-          return;
-        }
-        auto size = inputSizes[row];
-        auto offset = inputOffsets[row];
-        rawResultOffsets[row] = numSelected;
-        for (auto i = 0; i < size; ++i) {
-          if (!bitsDecoder.get()->isNullAt(offset + i) &&
-              bitsDecoder.get()->valueAt<bool>(offset + i)) {
-            ++rawResultSizes[row];
-            rawSelectedIndices[numSelected] = offset + i;
-            ++numSelected;
-          }
-        }
-      });
-    }
-
-    selectedIndices->setSize(numSelected * sizeof(vector_size_t));
-
-    return numSelected;
-  }
-};
-
-// See documentation at
-//    - https://prestodb.io/docs/current/functions/array.html
-//    - https://prestodb.io/docs/current/functions/lambda.html
-//    - https://prestodb.io/blog/2020/03/02/presto-lambda
-class ArrayFilterFunction : public FilterFunctionBase {
- public:
-  void apply(
-      const SelectivityVector& rows,
-      std::vector<VectorPtr>& args,
-      const TypePtr& /* outputType */,
-      exec::EvalCtx& context,
-      VectorPtr& result) const override {
-    VELOX_CHECK_EQ(args.size(), 2);
-    exec::LocalDecodedVector arrayDecoder(context, *args[0], rows);
-    auto& decodedArray = *arrayDecoder.get();
-
-    auto flatArray = flattenArray(rows, args[0], decodedArray);
-
-    VectorPtr elements = flatArray->elements();
-    BufferPtr resultSizes;
-    BufferPtr resultOffsets;
-    BufferPtr selectedIndices;
-    auto numSelected = doApply(
-        rows,
-        flatArray,
-        args[1],
-        {elements},
-        context,
-        resultOffsets,
-        resultSizes,
-        selectedIndices);
-
-    // Filter can pass along very large elements vectors that can hold onto
-    // memory and copy operations on them can further put memory pressure. We
-    // try to flatten them if the dictionary layer is much smaller than the
-    // elements vector.
-    auto wrappedElements = numSelected ? BaseVector::wrapInDictionary(
-                                             BufferPtr(nullptr),
-                                             std::move(selectedIndices),
-                                             numSelected,
-                                             std::move(elements),
-                                             true /*flattenIfRedundant*/)
-                                       : nullptr;
-    // Set nulls for rows not present in 'rows'.
-    BufferPtr newNulls = addNullsForUnselectedRows(flatArray, rows);
-    auto localResult = std::make_shared<ArrayVector>(
-        flatArray->pool(),
-        flatArray->type(),
-        std::move(newNulls),
-        rows.end(),
-        std::move(resultOffsets),
-        std::move(resultSizes),
-        wrappedElements);
-    context.moveOrCopyResult(localResult, rows, result);
-  }
-
-  static std::vector<std::shared_ptr<exec::FunctionSignature>> signatures() {
-    // array(T), function(T, boolean) -> array(T)
-    return {exec::FunctionSignatureBuilder()
-                .typeVariable("T")
-                .returnType("array(T)")
-                .argumentType("array(T)")
-                .argumentType("function(T,boolean)")
-                .build()};
-  }
-};
 
 // See documentation at
 //    - https://prestodb.io/docs/current/functions/map.html
@@ -440,9 +291,9 @@ class MapFilterFunction : public FilterFunctionBase {
 
 VELOX_DECLARE_VECTOR_FUNCTION_WITH_METADATA(
     udf_array_filter,
-    ArrayFilterFunction::signatures(),
+    ArrayFilterFunctionBase::signatures(),
     exec::VectorFunctionMetadataBuilder().defaultNullBehavior(false).build(),
-    std::make_unique<ArrayFilterFunction>());
+    std::make_unique<ArrayFilterFunctionBase>());
 
 VELOX_DECLARE_VECTOR_FUNCTION_WITH_METADATA(
     udf_map_filter,
