+++ velox/functions/sparksql/CMakeLists.txt
@@ -25,6 +25,7 @@ velox_add_library(
   DecimalArithmetic.cpp
   DecimalCeil.cpp
   DecimalCompare.cpp
+  FilterFunction.cpp
   Hash.cpp
   In.cpp
   LeastGreatest.cpp
@@ -34,6 +35,7 @@ velox_add_library(
   Size.cpp
   String.cpp
   TimestampUtils.cpp
+  Transform.cpp
   UnscaledValueFunction.cpp
 )
