+++ velox/functions/sparksql/FilterFunction.cpp
@@ -14,32 +14,31 @@
  * limitations under the License.
  */
 
-#include "velox/functions/lib/TransformFunctionBase.h"
+#include "velox/functions/lib/FilterFunctionBase.h"
 #include "velox/functions/sparksql/LambdaIndexUtil.h"
 
 namespace facebook::velox::functions::sparksql {
 namespace {
 
-/// Spark's transform function supports both signatures:
-/// 1. transform(array, x -> expr) - element only
-/// 2. transform(array, (x, i) -> expr) - element + index (Spark-specific)
+/// Spark's filter function supports both signatures:
+/// 1. filter(array, x -> predicate) - element only
+/// 2. filter(array, (x, i) -> predicate) - element + index (Spark-specific)
 ///
 /// See Spark documentation:
-/// https://spark.apache.org/docs/latest/api/sql/index.html#transform
-class TransformFunction : public TransformFunctionBase {
+/// https://spark.apache.org/docs/latest/api/sql/index.html#filter
+class FilterFunction : public ArrayFilterFunctionBase {
  public:
   /// Returns both base signature and Spark-specific signature with index.
   static std::vector<std::shared_ptr<exec::FunctionSignature>> signatures() {
-    auto sigs = TransformFunctionBase::signatures();
-    // Add Spark-specific signature: array(T), function(T, integer, U) ->
-    // array(U). Spark uses IntegerType (32-bit) for the index parameter.
+    auto sigs = ArrayFilterFunctionBase::signatures();
+    // Add Spark-specific signature: array(T), function(T, integer, boolean) ->
+    // array(T). Spark uses IntegerType (32-bit) for the index parameter.
     sigs.push_back(
         exec::FunctionSignatureBuilder()
             .typeVariable("T")
-            .typeVariable("U")
-            .returnType("array(U)")
+            .returnType("array(T)")
             .argumentType("array(T)")
-            .argumentType("function(T, integer, U)")
+            .argumentType("function(T,integer,boolean)")
             .build());
     return sigs;
   }
@@ -53,9 +52,8 @@ class TransformFunction : public TransformFunctionBase {
       exec::EvalCtx& context,
       std::vector<VectorPtr>& lambdaArgs) const override {
     // Check the lambda function type to see if it expects 2 input arguments.
-    // function(T, U) has 2 children (input T, output U) -> 1 input arg.
-    // function(T, integer, U) has 3 children (input T, index integer, output U)
-    // -> 2 input args.
+    // function(T, boolean) has 2 children (input T, output boolean) -> 1 arg.
+    // function(T, integer, boolean) has 3 children -> 2 args (element + index).
     if (args[1]->type()->size() == 3) {
       lambdaArgs.push_back(
           createIndexVector(flatArray, numElements, context.pool()));
@@ -66,9 +64,9 @@ class TransformFunction : public TransformFunctionBase {
 } // namespace
 
 VELOX_DECLARE_VECTOR_FUNCTION_WITH_METADATA(
-    udf_spark_transform,
-    TransformFunction::signatures(),
+    udf_spark_filter,
+    FilterFunction::signatures(),
     exec::VectorFunctionMetadataBuilder().defaultNullBehavior(false).build(),
-    std::make_unique<TransformFunction>());
+    std::make_unique<FilterFunction>());
 
 } // namespace facebook::velox::functions::sparksql
